Support repel re-use

+++ src/scrcmd.c
@@ -490,7 +490,7 @@ bool8 ScrCmd_additem(struct ScriptContext *ctx)
 bool8 ScrCmd_removeitem(struct ScriptContext *ctx)
 {
     u16 itemId = VarGet(ScriptReadHalfword(ctx));
-    u32 quantity = VarGet(ScriptReadHalfword(ctx));
+    u16 quantity = VarGet(ScriptReadHalfword(ctx));
 
     gSpecialVar_Result = RemoveBagItem(itemId, (u8)quantity);
     return FALSE;
@@ -499,7 +499,7 @@ bool8 ScrCmd_removeitem(struct ScriptContext *ctx)
 bool8 ScrCmd_checkitemspace(struct ScriptContext *ctx)
 {
     u16 itemId = VarGet(ScriptReadHalfword(ctx));
-    u32 quantity = VarGet(ScriptReadHalfword(ctx));
+    u16 quantity = VarGet(ScriptReadHalfword(ctx));
 
     gSpecialVar_Result = CheckBagHasSpace(itemId, (u8)quantity);
     return FALSE;
@@ -508,7 +508,7 @@ bool8 ScrCmd_checkitemspace(struct ScriptContext *ctx)
 bool8 ScrCmd_checkitem(struct ScriptContext *ctx)
 {
     u16 itemId = VarGet(ScriptReadHalfword(ctx));
-    u32 quantity = VarGet(ScriptReadHalfword(ctx));
+    u16 quantity = VarGet(ScriptReadHalfword(ctx));
 
     gSpecialVar_Result = CheckBagHasItem(itemId, (u8)quantity);
     return FALSE;
